ipa: libipa: agc: Change luminance target to piecewise linear function
Change the relative luminance target from a scalar value to a piecewise
linear function that needs to be sampled by the estimate lux value.

Also change the rkisp1 and ipu3 IPAs accordingly, as they use the libipa
agc. As they both don't yet have lux modules, hardcode them to a single
lux value for now.

This affects the format of the tuning files, but as there aren't yet any
this shouldn't be an issue.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>