utils: libtuning: modules: Add skeletal CCM module
Add a skeletal CCM module just so that we can have some CCM tuning
values that we can use to test during development of CCM in the IPAs. As
rkisp1 is the main target, we only add support for rkisp1 for now.

The parameters are mostly copied from the hardcoded values in ctt,
except for the metering modes.

As CCM is called CTK (for crosstalk) in rkisp1, that is the name that
will be used for rkisp1.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>