ipa: libipa: Add Vector class
Add a vector class to libipa. The original purpose of this is to replace
the floating-point Point class that Raspberry Pi used in their Pwl, as
that implementation of Point seemed more akin to a Vector than a Point.

This is added to libipa instead of to geometry.h to avoid public API
issues, plus this is not expected to be needed by applications.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>

---
Changes in v8:
- s/D/Rows/ in Vector class
- add z accessor function
- move SFINAE for x() and y() (and z()) to the functions instead of the
  class
- add vector to meson

Changes in v7:
- fix documentation
- fix license and copyright
- remove toString()

No change in v6

New in v5