ipa: simple: softisp: Extend to pass metadata
Extend the Simple IPA IPC to support returning a metadata controllist
when the process call has completed.

For efficiency, use the existing signal setIspParams() directly
to avoid having an extra async callback to return the metadata.

Merge the metadata reported by the ISP into any completing
request to provide to the application.

This should be further extended or improved to make use of the frame
context structures so that the metadata is tied to a specific frame and
request completion.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>