libcamera: rkisp1: Plumb the ConverterDW100 converter
Plumb the ConverterDW100 converter in the rkisp1 pipeline handler.
If the dewarper is found, it is instantiated and buffers are exported
from it, instead of RkISP1Path. Internal buffers are allocated for the
RkISP1Path in case where dewarper is going to be used.

The RKISP1 pipeline handler now supports scaler crop control through
ConverterDW100. Register the ScalerCrop control for the cameras created
in the RKISP1 pipeline handler.

Signed-off-by: Umang Jain <umang.jain@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>